+++ lib/alibaba_cloud_kit/oss/object.ex
@@ -152,6 +152,105 @@ defmodule AlibabaCloudKit.OSS.Object do
     }
   end
 
+  import AlibabaCloudKit.Utils,
+    only: [
+      utc_now: 1,
+      encode_json!: 1,
+      base64: 1,
+      hmac_sha1: 2
+    ]
+
+  @doc ~S"""
+  Presigns a `PostObject` operation with OSS V1 signature.
+
+  > OSS V1 signature is only supported at here.
+
+  > You should this function only for maintaining legacy code. In other case,
+  > please use `presign_post_object/3`.
+
+  ## Examples
+
+      region = "oss-us-west-1"
+      bucket = "example-bucket"
+
+      key = "lenna.png"
+      acl = "private"
+      max_size_in_bytes = 1024 * 1024 * 5
+
+      conditions = [
+        ["eq", "$key", key],
+        ["eq", "$x-oss-object-acl", acl],
+        ["content-length-range", 1, max_size_in_bytes]
+      ]
+
+      seconds_to_expire = 3600
+
+      opts = [
+        access_key_id: "...",
+        access_key_secret: "...",
+        region: region,
+        bucket: bucket
+      ]
+
+      %{
+        policy: policy,
+        signature: signature
+      } = AlibabaCloudKit.OSS.presign_post_object_v1(conditions, seconds_to_expire, opts)
+
+  To use the returned data, you should built a form with multipart data, and send it:
+
+      alias Tesla.Multipart
+
+      url = "https://#{bucket}.#{region}.aliyuncs.com"
+
+      multipart =
+        Multipart.new()
+        |> Multipart.add_field("key", key)
+        |> Multipart.add_field("x-oss-object-acl", acl)
+        |> Multipart.add_field("OSSAccessKeyId", access_key_id)
+        |> Multipart.add_field("policy", policy)
+        |> Multipart.add_field("Signature", signature)
+        |> Multipart.add_file_content(binary_of_file, "lenna.png")
+
+      {:ok, %{status: 204}} = Tesla.post(url, multipart)
+
+  """
+  @spec presign_post_object_v1(list(), pos_integer(), sign_opts()) :: %{
+          policy: String.t(),
+          signature: String.t()
+        }
+  def presign_post_object_v1(conditions, seconds_to_expire, opts)
+      when is_list(conditions) and is_integer(seconds_to_expire) do
+    %{
+      access_key_secret: access_key_secret,
+      bucket: bucket,
+      at: at
+    } =
+      opts
+      |> NimbleOptions.validate!(@sign_opts_definition)
+      |> Map.new()
+
+    at = at || utc_now(:second)
+
+    policy =
+      %{
+        expiration: build_expiration(at, seconds_to_expire),
+        conditions: [
+          %{bucket: bucket}
+          | conditions
+        ]
+      }
+      |> encode_json!()
+      |> base64()
+
+    signature = access_key_secret |> hmac_sha1(policy) |> base64()
+
+    %{
+      policy: policy,
+      signature: signature
+    }
+  end
+
   defp build_expiration(%DateTime{} = start, seconds_to_expire) do
     start
     |> DateTime.add(seconds_to_expire, :second)