+++ test/alibaba_cloud_kit/oss/object_test.exs
@@ -14,9 +14,10 @@ defmodule AlibabaCloudKit.OSS.ObjectTest do
 
   # Because the internal of JSON encoder, encoding a map won't always produce
   # the same string on different platforms, which will cause the test to fail.
-  # To make the CI pass, I tag this test, and ignore it from normal flow.
+  # To make the CI pass, I tag these tests, and ignore them from normal flow.
+
   @tag uncertain: true
-  test "presign_post_object/5", %{
+  test "presign_post_object/3", %{
     access_key_id: access_key_id,
     access_key_secret: access_key_secret,
     region: region,
@@ -51,5 +52,39 @@ defmodule AlibabaCloudKit.OSS.ObjectTest do
                "43fb9eb04136cdaaab110bfb33c4c7daba4beebdad63889e7f3a460242596d04",
              "x-oss-signature-version": "OSS4-HMAC-SHA256"
            } == Object.presign_post_object(conditions, seconds_to_expire, opts)
+  end
+
+  @tag uncertain: true
+  test "presign_post_object_v1/3", %{
+    access_key_id: access_key_id,
+    access_key_secret: access_key_secret,
+    region: region,
+    bucket: bucket
+  } do
+    key = "object/presign_post_object.png"
+    acl = "private"
+    max_size_in_bytes = 1024 * 1024 * 100
+
+    conditions = [
+      ["eq", "$key", key],
+      ["eq", "$x-oss-object-acl", acl],
+      ["content-length-range", 1, max_size_in_bytes]
+    ]
+
+    seconds_to_expire = 3600
+
+    opts = [
+      access_key_id: access_key_id,
+      access_key_secret: access_key_secret,
+      region: region,
+      bucket: bucket,
+      at: ~U[2024-07-13 13:34:14Z]
+    ]
+
+    assert %{
+             signature: "rOLRLFOeZRsJ+8WA6wAiPLCda14=",
+             policy:
+               "eyJjb25kaXRpb25zIjpbeyJidWNrZXQiOiJleGFtcGxlLWJ1Y2tldCJ9LFsiZXEiLCIka2V5Iiwib2JqZWN0L3ByZXNpZ25fcG9zdF9vYmplY3QucG5nIl0sWyJlcSIsIiR4LW9zcy1vYmplY3QtYWNsIiwicHJpdmF0ZSJdLFsiY29udGVudC1sZW5ndGgtcmFuZ2UiLDEsMTA0ODU3NjAwXV0sImV4cGlyYXRpb24iOiIyMDI0LTA3LTEzVDE0OjM0OjE0WiJ9"
+           } == Object.presign_post_object_v1(conditions, seconds_to_expire, opts)
   end
 end